+++ flashservices/src/main/java/com/inalab/dao/CommonDao.java
@@ -4,13 +4,13 @@ import java.util.List;
 
 public interface CommonDao<T> {
 
-	int insert(T record, String sql);
+	long insert(T record, String sql);
 
-	T getById(int id, String sql);
+	T getById(long id, String sql);
 
 	boolean update(T record, String sql);
 	
-	List<T> getListById(int id, String sql);
+	List<T> getListById(long id, String sql);
 	
 
 }
